Boyscouting and refactoring
