+++ lib/cli/generators/ANGULAR/index.js
@@ -1,14 +1,16 @@
 import mergeDirs from 'merge-dirs';
 import path from 'path';
-import { editStorybookTsConfig } from './angular-helpers';
+import {
+  editStorybookTsConfig,
+  getAngularAppTsConfigJson,
+  writeAngularAppTsConfig,
+} from './angular-helpers';
 import {
   getVersions,
   getPackageJson,
   writePackageJson,
   getBabelDependencies,
   installDependencies,
-  getAngularAppTsConfigJson,
-  writeAngularAppTsConfig,
 } from '../../lib/helpers';
 
 async function addDependencies(npmOptions) {
