+++ lib/cli/lib/helpers.js
@@ -62,44 +62,25 @@ export function getBowerJson() {
   return JSON.parse(jsonContent);
 }
 
-export function getAngularJson() {
-  const angularJsonPath = path.resolve('angular.json');
-  if (!fs.existsSync(angularJsonPath)) {
+export function readFileAsJson(jsonPath) {
+  const filePath = path.resolve(jsonPath);
+  if (!fs.existsSync(filePath)) {
     return false;
   }
 
-  const jsonContent = fs.readFileSync(angularJsonPath, 'utf8');
+  const jsonContent = fs.readFileSync(filePath, 'utf8');
   return JSON.parse(jsonContent);
 }
 
-export function getAngularAppTsConfigPath() {
-  const angularJson = getAngularJson();
-  const { defaultProject } = angularJson;
-  const tsConfigPath = angularJson.projects[defaultProject].architect.build.options.tsConfig;
-
-  if (!tsConfigPath || !fs.existsSync(path.resolve(tsConfigPath))) {
-    return false;
-  }
-  return tsConfigPath;
-}
-
-export function getAngularAppTsConfigJson() {
-  const tsConfigPath = getAngularAppTsConfigPath();
-
-  if (!tsConfigPath || !fs.existsSync(path.resolve(tsConfigPath))) {
+export function writeFileAsJson(jsonPath, content) {
+  const filePath = path.resolve(jsonPath);
+  if (!fs.existsSync(filePath)) {
     return false;
   }
 
-  const jsonContent = fs.readFileSync(tsConfigPath, 'utf8');
-  return JSON.parse(jsonContent);
-}
-
-export function writeAngularAppTsConfig(tsConfigJson) {
-  const content = `${JSON.stringify(tsConfigJson, null, 2)}\n`;
-  const tsConfigPath = getAngularAppTsConfigPath();
-  if (tsConfigPath) {
-    fs.writeFileSync(path.resolve(tsConfigPath), content, 'utf8');
-  }
+  const jsonContent = fs.readFileSync(filePath, 'utf8');
+  fs.writeFileSync(filePath, JSON.stringify(content, null, 2));
+  return true;
 }
 
 export function writePackageJson(packageJson) {
